# portal/backend/openapi/portal-backend.yaml
openapi: 3.0.3
info:
  title: OpenFGC Portal Backend API
  version: 0.3.0
  description: >-
    Stateless portal backend endpoints, including health checks, portal-facing
    user routes, and explicit passthrough routes to Consent Management API v0.3.0.
    For proxied endpoints in this spec, response payload shape is inherited from the
    consent-management API.
servers:
  - url: http://localhost:8080
security:
  - PortalOAuthDocumentation: []
    AccessTokenPart2: []
...
paths:
...
  /api/consents:
    get:
      tags: [Passthrough]
      summary: Search consents
      security:
        - PortalOAuthDocumentation: [portal:consents:read:any]
          AccessTokenPart2: []
      responses:
        '200': { description: Upstream response }
    post:
      tags: [Passthrough]
      summary: Create consent
      security:
        - PortalOAuthDocumentation: [portal:consents:write:any]
          AccessTokenPart2: []
      responses:
        '200': { description: Upstream response }
...
components:
  securitySchemes:
    PortalOAuthDocumentation:
      type: oauth2
      x-documentation-only: true
      description: >-
        Documentation model for the OAuth2 scopes enforced by the BFF. The URLs
        are deployment-specific placeholders. Note that the Swagger UI cannot execute the
        actual split-token flow. The Authorization header contains only the
        JavaScript-readable first half of the access token, and the BFF combines
        it with AccessTokenPart2 before validation.
      flows:
        authorizationCode:
          authorizationUrl: https://idp.example.com/oauth2/authorize
          tokenUrl: https://idp.example.com/oauth2/token
          scopes:
            'portal:consents:read:self': Read the current user's consents.
            'portal:consents:write:self': Approve or revoke the current user's consents.
            'portal:consents:read:any': Read consents through passthrough APIs.
            'portal:consents:write:any': Modify consents through passthrough APIs.
            'portal:elements:read': Read consent elements.
            'portal:elements:write': Modify consent elements.
            'portal:purposes:read': Read consent purposes.
            'portal:purposes:write': Modify consent purposes.
    AccessTokenPart2:
      type: apiKey
      in: cookie
      name: portal-at-p2
      description: Configurable HttpOnly second-half access-token cookie.
...